Apply patch #1157756 from Jeremy Thomerson:
This is to address feature request 1151779. It adds a
<package> tag to the XML coverage report. It has line-
rate and branch-rate as the summary attributes. If these
need to be formatted differently (as elements?) or named
differently, it should be easy to do.

I moved classes like Clazz, Package, Coverage, etc...
up from reporting.html ro reporting so that they could be
used in both XMLReport and HTMLReport.

I added the raw CoverageData instance to Clazz,
because it looked like the XMLReport needed things that
the HTMLReport didn't, and it was pulling this from the
CoverageData instance. (NOTE: we may want to
eliminate a lot of the method implementations in Clazz,
and let them defer to the raw CoverageData. It seems
like we may be implementing the same thing multiple
places)


I also fixed a bug where the per-method branch and line
coverage rates in the XML report were always "NaN"