Add a test which will validate the generated coverage.xml file.  It currently only checks that the XML is valid, but I am writing a scheme so so it will be able to validate the structure, as well.
Let me know if I need to check in any of the other xerces jars--
I seem to be able to run everything ok with just this one.

Also, if anyone wants to convert the CREATION of the
XMLReport to use xerces that's fine with me.  But I don't
mind leaving it as-is.