Patch [1543280] - Adding cobertura.properties file for use in limited environments.
When working in limited environments. (ones that cannot set a System Property, or will not propogate a System Property) the ability to specify the datafile location is impossible.

To correct this, the attached patch will utilize a cobertura.properties file located in the classpath to properly set the net.sourceforge.cobertura.datafile property.

This patch introduces a ConfigurationUtil that returns a value after looking for a System Property first, then a cobertura.properties key, then a default value.

This patch also addresses a situation where the output location for the datafile is a deep directory that doesn't exist (yet).