Add filtering of command-line scripts
The command-line scripts were modified to include tokens for the
versions of the different libraries, which will change over time and
were out of date. The modified POM now filters the scripts,
substituting in the versions as specified in the POM, and the modified
assembly descriptor includes them in the distribution.